+++ src/app/experiments/[slug]/page.tsx
@@ -1,14 +1,47 @@
 import SingleExperimentContent from "@/contents/single-experiment";
+import { getAllSlugs, getExperimentBySlug, getRelatedExperiments } from "@/lib/helpers/experiments";
+import { createMetaAlternates } from "@/lib/utils";
+import { Metadata } from "next";
+import { notFound } from "next/navigation";
 
-interface SingleExperimentPage{
+interface SingleExperimentPageProps{
   params: Promise<{slug: string}>
 }
-
 export const revalidate = 86400
 
-export default async function SingleExperimentPage({params}: SingleExperimentPage){
+export const generateStaticParams = async() => {
+  const allSlugs = await getAllSlugs();
+  return allSlugs.map(slug => ({slug}))
+}
+
+export const generateMetadata = async({params}:SingleExperimentPageProps): Promise<Metadata> => {
+     const {slug} = await params;
+     const currExperiment = await getExperimentBySlug(slug);
+     if(!currExperiment) return notFound();
+     return {
+          title: currExperiment.title,
+          description: currExperiment.description,
+          keywords: currExperiment.tags,
+          authors: [
+               {
+                    name: "Arsen G.",
+                    url: "https://arsen-2005.vercel.app/"
+               }
+          ],
+          alternates: createMetaAlternates(`/experiments/${slug}`)
+     }
+}
+
+export default async function SingleExperimentPage({params}: SingleExperimentPageProps){
      const {slug} = await params
+     const currExperiment = await getExperimentBySlug(slug);
+     if(!currExperiment) notFound();
+     const relatedExperiments = await getRelatedExperiments(slug, currExperiment.tags)
      return (
-          <SingleExperimentContent slug={slug}/>
+          <SingleExperimentContent
+               experiment={currExperiment}
+               related={relatedExperiments}
+               slug={slug}
+          />
      )
 }