+++ changelog.md
@@ -6,11 +6,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ---
 ## 0.3.0 - Unreleased
 ### Added
-- Empty component for experiments as a fallback state
-### Initialized
 - MDX support with LATEX plugin
 - MDX content example
 - MDX Helper functions
+- Experiments Support
+- Empty component for experiments as a fallback state
+### Changed
+- Added working pagination and filtering on the experiments page
 
 ## 0.2.0 - 2026-09-08
 ### Added